# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Learning System API",
    "description": "Backend API for Online Learning Platform - Challenges 1-3\n\n## Challenges Completed:\n- **Challenge 1**: Database Design (PostgreSQL, 7 tables, 3NF normalized)\n- **Challenge 2**: CRUD Operations (Courses & Classes)\n- **Challenge 3**: Authentication (Register, Login, JWT, Profile)\n\n## Features:\n- ✅ User Registration & Login with JWT\n- ✅ Role-based Authorization (admin, instructor, student)\n- ✅ Course Management (CRUD with instructor/admin)\n- ✅ Class Management (CRUD with pagination)\n- ✅ Pagination & Filtering\n- ✅ Error Handling & Validation\n- ✅ Logging & Monitoring\n\n## Test Credentials:\n- **Email**: student1@example.com\n- **Password**: password123\n- **Role**: student",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register New User",
        "description": "Register a new user account. User will have 'student' role by default.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "password", "confirmPassword"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "newstudent@example.com"
                  },
                  "name": {
                    "type": "string",
                    "example": "New Student"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 6,
                    "example": "password123"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": ["inactive", "active"]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or email already exists"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login User",
        "description": "Login and get JWT token. Token valid for 1 hour.",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "student1@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string",
                          "description": "JWT token for Authorization header"
                        },
                        "expires_in": {
                          "type": "string",
                          "example": "1h"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "email": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": ["admin", "instructor", "student"]
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password"
          }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get User Profile",
        "description": "Get current user profile (requires JWT token)",
        "operationId": "getProfile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid token"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token received from /api/auth/login endpoint"
      }
    }
  }
}